src/test/java/medium/graph/Q0787CheapestFlightWithKStops/SolutionTest.java
package medium.graph.Q0787CheapestFlightWithKStops;

import org.junit.jupiter.api.Test;

import static org.junit.jupiter.api.Assertions.*;

class SolutionTest {
    @Test
    void testOne() {
        assertEquals(700, new Solution().findCheapestPrice(4, new int[][]{{0,1,100},{1,2,100},{2,0,100},{1,3,600},{2,3,200}}, 0, 3, 1));
    }

    @Test
    void testTwo() {
        assertEquals(200, new Solution().findCheapestPrice(3, new int[][]{{0,1,100},{1,2,100},{0,2,500}}, 0, 2, 1));
    }

    @Test
    void testThree() {
        assertEquals(500, new Solution().findCheapestPrice(3, new int[][]{{0,1,100},{1,2,100},{0,2,500}}, 0, 2, 0));
    }
}